# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arth API",
    "description": "Realtime paper-trading broker API (SSE market stream, session cookie auth).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://127.0.0.1:4000", "description": "Local" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arth_session"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "cash": { "type": "number" },
          "kycComplete": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "phone", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created; session cookie set" },
          "400": { "description": "Validation error" },
          "409": { "description": "Email taken" },
          "429": { "description": "Rate limited" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK; session cookie set" },
          "401": { "description": "Invalid credentials" },
          "429": { "description": "Rate limited" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Current user",
        "tags": ["Auth"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated" }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "summary": "List orders",
        "tags": ["Orders"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Orders for current user" },
          "401": { "description": "Not authenticated" },
          "429": { "description": "Rate limited" }
        }
      },
      "post": {
        "summary": "Place order",
        "tags": ["Orders"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symbol": { "type": "string" },
                  "side": { "type": "string", "enum": ["BUY", "SELL"] },
                  "qty": { "type": "number" },
                  "type": { "type": "string" },
                  "price": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order accepted" },
          "201": { "description": "Order created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Not authenticated" },
          "429": { "description": "Rate limited" }
        }
      }
    },
...
    "/api/learn/content": {
      "get": {
        "summary": "Learn hub content + progress",
        "tags": ["Learn"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Lessons and challenges" }
        }
      }
    },
    "/api/learn/mode": {
      "patch": {
        "summary": "Set learn / paper mode",
        "tags": ["Learn"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Updated" }
        }
      }
    },
    "/api/conditional": {
      "get": {
        "summary": "List conditional orders",
        "tags": ["Conditional"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Conditional orders" }
        }
      },
      "post": {
        "summary": "Create conditional order",
        "tags": ["Conditional"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Created" },
          "201": { "description": "Created" }
        }
      }
    },
    "/api/conditional/{id}": {
      "delete": {
        "summary": "Cancel conditional order",
        "tags": ["Conditional"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" }
        }
      }
    },
...
  }
}